+++ crates/mako/src/module.rs
@@ -296,7 +296,13 @@ impl Module {
 
                 for n in script.ast.body.iter() {
                     match n.as_stmt() {
-                        None => return Err(anyhow!("Error: {:?} not a stmt in ", self.id.id)),
+                        None => {
+                            return Err(anyhow!(
+                                "Error: not a stmt found in {:?}, ast: {:?}",
+                                self.id.id,
+                                n,
+                            ));
+                        }
                         Some(stmt) => {
                             stmts.push(stmt.clone());
                         }