+++ crates/mako/src/transformers/transform_dep_replacer.rs
@@ -26,7 +26,9 @@ pub struct DepReplacer<'a> {
 
 #[derive(Debug, Clone)]
 pub struct DependenciesToReplace {
-    pub resolved: HashMap<String, String>,
+    // resolved stores the "source" maps to (generate_id, raw_module_id)
+    // e.g. "react" => ("hashed_id", "/abs/to/react/index.js")
+    pub resolved: HashMap<String, (String, String)>,
     pub missing: HashMap<String, Dependency>,
     pub ignored: Vec<String>,
 }
@@ -102,19 +104,28 @@ impl VisitMut for DepReplacer<'_> {
                                 spread: None,
                                 expr: Box::new(miss_throw_stmt(&source_string)),
                             };
+                            return;
                         }
                         None => {
                             self.replace_source(source);
                         }
                     }
 
-                    let file_request = parse_path(&source_string).unwrap();
-                    if is_css_path(&file_request.path)
-                        && (file_request.query.is_empty() || file_request.has_query("modules"))
+                    let is_dep_replaceable = if let Some((_, raw_id)) =
+                        self.to_replace.resolved.get(&source_string)
                     {
+                        let file_request = parse_path(raw_id).unwrap();
+                        is_css_path(&file_request.path)
+                            && (file_request.query.is_empty() || file_request.has_query("modules"))
+                    } else {
+                        false
+                    };
+
+                    if is_dep_replaceable {
                         // remove `require('./xxx.css');`
                         if is_commonjs_require_flag {
-                            *expr = Expr::Lit(quote_str!("").into())
+                            *expr = Expr::Lit(quote_str!("").into());
+                            return;
                         } else {
                             // `import('./xxx.css')` 中的 css 模块会被拆分到单独的 chunk, 这里需要改为加载 css chunk
                             let module_graph = self.context.module_graph.read().unwrap();
@@ -131,11 +142,14 @@ impl VisitMut for DepReplacer<'_> {
                                     // `import('./xxx.css')` => `__mako_require__.ensure('./xxx.css')`
                                     *expr = member_expr!(DUMMY_SP, __mako_require__.ensure)
                                         .as_call(DUMMY_SP, vec![quote_str!(chunk_id).as_arg()]);
+                                    return;
                                 } else {
-                                    *expr = Expr::Lit(quote_str!("").into())
+                                    *expr = Expr::Lit(quote_str!("").into());
+                                    return;
                                 }
                             } else {
-                                *expr = Expr::Lit(quote_str!("").into())
+                                *expr = Expr::Lit(quote_str!("").into());
+                                return;
                             }
                         }
                     }
@@ -168,7 +182,7 @@ impl DepReplacer<'_> {
     fn replace_source(&mut self, source: &mut Str) {
         let to_replace =
             if let Some(replacement) = self.to_replace.resolved.get(&source.value.to_string()) {
-                replacement.clone()
+                replacement.0.clone()
             } else if self.to_replace.ignored.contains(&source.value.to_string()) {
                 "$$IGNORED$$".to_string()
             } else {
@@ -230,7 +244,7 @@ mod tests {
     use crate::assert_display_snapshot;
     use crate::ast::build_js_ast;
     use crate::compiler::Context;
-    use crate::module::{Dependency, ModuleId, ResolveType};
+    use crate::module::{Dependency, Module, ModuleId, ResolveType};
     use crate::test_helper::transform_ast_with;
     use crate::transformers::test_helper::transform_js_code;
     use crate::transformers::transform_dep_replacer::{DepReplacer, DependenciesToReplace};
@@ -240,17 +254,29 @@ mod tests {
         let context: Arc<Context> = Arc::new(Default::default());
 
         GLOBALS.set(&context.meta.script.globals, || {
-            let mut ast = build_js_ast("index.jsx",
-                                       r#"require("react")"#
-                                       , &context.clone()).unwrap();
+            let mut ast =
+                build_js_ast("index.jsx", r#"require("react")"#, &context.clone()).unwrap();
 
-            let to_replace  = DependenciesToReplace {
-                resolved: hashmap! {"react".to_string()=> "/root/node_modules/react/index.js".to_string()},
+            let to_replace = DependenciesToReplace {
+                resolved: hashmap! {"react".to_string()=>
+                    (
+                        "/root/node_modules/react/index.js".to_string(),
+                        "/root/node_modules/react/index.js".to_string()
+                    )
+                },
                 missing: HashMap::new(),
                 ignored: vec![],
             };
 
             let cloned = context.clone();
+
+            cloned.module_graph.write().unwrap().add_module(Module {
+                id: "index.jsx".to_string().into(),
+                is_entry: false,
+                info: None,
+                side_effects: false,
+            });
+
             let module_id = ModuleId::new("index.jsx".to_string());
             let mut visitor: Box<dyn VisitMut> = Box::new(chain!(
                 resolver(ast.unresolved_mark, ast.top_level_mark, false),
@@ -263,7 +289,11 @@ mod tests {
                 }
             ));
 
-            assert_display_snapshot!(transform_ast_with(&mut ast.ast, &mut visitor, &context.meta.script.cm));
+            assert_display_snapshot!(transform_ast_with(
+                &mut ast.ast,
+                &mut visitor,
+                &context.meta.script.cm
+            ));
         });
     }
 
@@ -370,14 +400,20 @@ mod tests {
     }
 
     fn transform_code(code: &str) -> String {
-        let context = Arc::new(Default::default());
+        let context: Arc<Context> = Arc::new(Default::default());
         let unresolved_mark = Default::default();
         let top_level_mark = Default::default();
+
         let mut visitor = DepReplacer {
             module_id: &ModuleId::new("index.jsx".into()),
             to_replace: &DependenciesToReplace {
                 resolved: hashmap! {
-                    "x".to_string() => "/x/index.js".to_string()
+                    "x".to_string() =>
+                    (
+
+                     "/x/index.js".to_string(),
+                     "/x/index.js".to_string()
+                    )
                 },
                 missing: hashmap! {},
                 ignored: vec![],