+++ crates/node/src/lib.rs
@@ -3,39 +3,24 @@
 #[macro_use]
 extern crate napi_derive;
 
-use std::sync::mpsc::Sender;
 use std::sync::{Arc, Once};
-use std::time::UNIX_EPOCH;
 
+use js_plugin::JsPlugin;
 use mako::compiler::{Args, Compiler};
 use mako::config::Config;
-use mako::dev::{DevServer, OnDevCompleteParams, Stats};
+use mako::dev::DevServer;
 use mako::logger::init_logger;
 use mako::plugin::Plugin;
 use napi::bindgen_prelude::*;
-use napi::{JsObject, JsString, JsUnknown, NapiRaw, Status};
+use napi::{JsObject, Status};
+use tsfn::{JsHooks, TsFnHooks};
 
-use crate::plugin_write::create_fs_write_plugin;
-
-mod plugin_less;
-mod plugin_write;
+mod js_plugin;
+mod tsfn;
 
 pub(crate) mod threadsafe_function;
 
 static LOG_INIT: Once = Once::new();
-
-#[napi(object)]
-pub struct JsHooks {
-    #[napi(ts_type = "(filePath: string) => Promise<string> ;")]
-    pub on_compile_less: Option<JsFunction>,
-    #[napi(ts_type = "(data: {isFirstCompile: boolean; time: number; stats: {
-        startTime: number;
-        endTime: number;
-    }}) =>void ;")]
-    pub on_build_complete: Option<JsFunction>,
-    #[napi(ts_type = "(path: string, content: Buffer) => Promise<void>;")]
-    pub on_generate_file: Option<JsFunction>,
-}
 
 #[napi(object)]
 pub struct BuildParams {
@@ -118,6 +103,12 @@ pub struct BuildParams {
     optimization?: false | {
         skipModules?: boolean;
     };
+    react?: {
+        runtime?: "automatic" | "classic";
+        pragma?: string;
+        importSource?: string;
+        pragmaFrag?: string;
+    };
 }"#)]
     pub config: serde_json::Value,
     pub hooks: JsHooks,
@@ -130,59 +121,10 @@ pub fn build(env: Env, build_params: BuildParams) -> napi::Result<JsObject> {
         init_logger();
     });
 
-    let on_build_complete = if let Some(on_build_complete) = build_params.hooks.on_build_complete {
-        let func = threadsafe_function::ThreadsafeFunction::create(
-            env.raw(),
-            unsafe { on_build_complete.raw() },
-            0,
-            |ctx: threadsafe_function::ThreadSafeCallContext<OnDevCompleteParams>| {
-                let mut obj = ctx.env.create_object()?;
-                let mut stats = ctx.env.create_object()?;
-                stats.set_named_property(
-                    "startTime",
-                    ctx.env.create_int64(ctx.value.stats.start_time as i64)?,
-                )?;
-                stats.set_named_property(
-                    "endTime",
-                    ctx.env.create_int64(ctx.value.stats.end_time as i64)?,
-                )?;
-                obj.set_named_property("isFirstCompile", ctx.value.is_first_compile)?;
-                obj.set_named_property("time", ctx.env.create_int64(ctx.value.time as i64))?;
-                obj.set_named_property("stats", stats)?;
-                ctx.callback.unwrap().call(None, &[obj])?;
-                Ok(())
-            },
-        )?;
-        Some(func)
-    } else {
-        None
-    };
-
-    let less_plugin = if let Some(on_compile_less) = build_params.hooks.on_compile_less {
-        let on_compile_less = threadsafe_function::ThreadsafeFunction::create(
-            env.raw(),
-            unsafe { on_compile_less.raw() },
-            0,
-            |ctx: threadsafe_function::ThreadSafeCallContext<ReadMessage>| {
-                let str = ctx.env.create_string(&ctx.value.message)?;
-                let result = ctx.callback.unwrap().call(None, &[str])?;
-                await_promise(ctx.env, result, ctx.value.tx).unwrap();
-                Ok(())
-            },
-        )?;
-        Some(Arc::new(plugin_less::LessPlugin { on_compile_less }))
-    } else {
-        None
-    };
-
     let mut plugins: Vec<Arc<dyn Plugin>> = vec![];
-    if let Some(less_plugin) = less_plugin {
-        plugins.push(less_plugin);
-    }
-
-    if let Some(plugin) = create_fs_write_plugin(&env, build_params.hooks.on_generate_file) {
-        plugins.push(Arc::new(plugin));
-    }
+    let tsfn_hooks = TsFnHooks::new(env, &build_params.hooks);
+    let plugin = JsPlugin { hooks: tsfn_hooks };
+    plugins.push(Arc::new(plugin));
 
     let root = std::path::PathBuf::from(&build_params.root);
     let default_config = serde_json::to_string(&build_params.config).unwrap();
@@ -194,8 +136,6 @@ pub fn build(env: Env, build_params: BuildParams) -> napi::Result<JsObject> {
         let (deferred, promise) = env.create_deferred()?;
         env.execute_tokio_future(
             async move {
-                let start_time = std::time::SystemTime::now();
-
                 let compiler =
                     Compiler::new(config, root.clone(), Args { watch: true }, Some(plugins))
                         .map_err(|e| napi::Error::new(Status::GenericFailure, format!("{}", e)));
@@ -212,24 +152,9 @@ pub fn build(env: Env, build_params: BuildParams) -> napi::Result<JsObject> {
                     deferred.reject(e);
                     return Ok(());
                 }
-
-                let end_time = std::time::SystemTime::now();
-                let params = OnDevCompleteParams {
-                    is_first_compile: true,
-                    time: end_time.duration_since(start_time).unwrap().as_millis() as u64,
-                    stats: Stats {
-                        start_time: start_time.duration_since(UNIX_EPOCH).unwrap().as_millis()
-                            as u64,
-                        end_time: end_time.duration_since(UNIX_EPOCH).unwrap().as_millis() as u64,
-                    },
-                };
-                call_on_build_complete(&on_build_complete, params);
                 let d = DevServer::new(root.clone(), Arc::new(compiler));
                 deferred.resolve(move |env| env.get_undefined());
-                d.serve(move |params| {
-                    call_on_build_complete(&on_build_complete, params);
-                })
-                .await;
+                d.serve(move |_params| {}).await;
                 Ok(())
             },
             move |&mut _, _res| Ok(()),
@@ -259,54 +184,4 @@ pub fn build(env: Env, build_params: BuildParams) -> napi::Result<JsObject> {
         });
         Ok(promise)
     }
-}
-
-fn call_on_build_complete(
-    on_build_complete: &Option<threadsafe_function::ThreadsafeFunction<OnDevCompleteParams>>,
-    params: OnDevCompleteParams,
-) {
-    if let Some(func) = on_build_complete {
-        func.call(
-            params,
-            threadsafe_function::ThreadsafeFunctionCallMode::NonBlocking,
-        );
-    }
-}
-
-fn await_promise(
-    env: Env,
-    result: JsUnknown,
-    tx: Sender<napi::Result<String>>,
-) -> napi::Result<()> {
-    // If the result is a promise, wait for it to resolve, and send the result to the channel.
-    // Otherwise, send the result immediately.
-    if result.is_promise()? {
-        let result: JsObject = result.try_into()?;
-        let then: JsFunction = result.get_named_property("then")?;
-        let tx2 = tx.clone();
-        let cb = env.create_function_from_closure("callback", move |ctx| {
-            let res = ctx.get::<JsString>(0)?.into_utf8()?;
-            let s = res.into_owned()?;
-            tx.send(Ok(s)).unwrap();
-            ctx.env.get_undefined()
-        })?;
-        let eb = env.create_function_from_closure("error_callback", move |ctx| {
-            let res = ctx.get::<JsUnknown>(0)?;
-            tx2.send(Err(napi::Error::from(res))).unwrap();
-            ctx.env.get_undefined()
-        })?;
-        then.call(Some(&result), &[cb, eb])?;
-    } else {
-        let result: JsString = result.try_into()?;
-        let utf8 = result.into_utf8()?;
-        let s = utf8.into_owned()?;
-        tx.send(Ok(s)).unwrap();
-    }
-
-    Ok(())
-}
-
-pub struct ReadMessage {
-    pub message: String,
-    pub tx: Sender<Result<String>>,
 }