+++ e2e/fixtures.umi/config.less.modifyVars/.umirc.ts
@@ -1,0 +1,11 @@
+export default {
+  mfsu: false,
+  theme: {
+    '@primary-color': 'red',
+  },
+  lessLoader: {
+    modifyVars: {
+      '@primary-color': 'blue',
+    },
+  },
+};